requirements.txt
amqp==5.0.9
asgiref==3.4.1
autopep8==1.6.0
billiard==3.6.4.0
celery==5.2.3
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coreapi==2.3.3
coreschema==0.0.4
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.8
django-cors-headers==3.10.0
django-filter==21.1
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
kombu==5.2.3
MarkupSafe==2.0.1
packaging==21.3
Pillow==8.4.0
prompt-toolkit==3.0.24
psycopg2-binary==2.9.1
pycodestyle==2.8.0
PyJWT==2.3.0
pyparsing==3.0.1
pytz==2021.3
redis==4.1.0
requests==2.26.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
wrapt==1.13.3
